Hoist mock OIDC server into testutil package
Move DeployParameterizedOIDCServer and its Python script from the
virtualmcp helpers into the shared testutil package so acceptance
tests can reuse it for per-user rate limiting E2E tests.

The virtualmcp package retains a thin wrapper that delegates to
testutil for backwards compatibility.

Part of #4550
